pyproject.toml
[project]
name = "online-school"
version = "0.1.0"
description = ""
authors = [
    {name = "Ксения",email = "kssshaab@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0,<7.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "django-filter (>=25.2,<26.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "coverage (>=7.13.1,<8.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
black = "^25.12.0"
flake8 = "^7.3.0"
isort = "^7.0.0"

[tool.black]
line-length = 119
exclude = '''
/(
    \.venv
  | venv
  | \.git
  | \.mypy_cache
  | \.pytest_cache
  | build
  | dist
)/
'''

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv|README\.md'
